TG: ODSliderItem - use double for now, have to think about a serious fix to support different data types; test.menu - wind speed slider

+++ AppsV2/OceanDemo/Menu/ODSliderItem.m
@@ -165,8 +165,8 @@
 
     if ( target != nil )
     {
-        float currentValue;
-        float normalisedValue;
+        double currentValue;
+        double normalisedValue;
 
         ODObjCGetVariable(target, offset, size, &currentValue);
         normalisedValue = (currentValue - minimumValue) / (maximumValue - minimumValue);
@@ -219,7 +219,7 @@
 
     if ( target != nil )
     {
-        const float scaledValue = [ self scaledValue ];
+        const double scaledValue = [ self scaledValue ];
         ODObjCSetVariable(target, offset, size, &scaledValue);
     }
 }
